Add first node script.
This introduces the first node type block, which pulls the node script into Zazu
to be executed multiple times later.

There are docs for both the Prefix Node Script and the PluginContext that gets
pushed into the script.

There is initial logging for all the usual things, as well as logging for if
loading the script fails.