Correctly handle reseting query and default queries.
When we open zazu, we want to send an empty query to all plugins, which works
great for scoped plugins.

When we hide zazu, we use to send an empty query too, assuming that would
"clear" the results, but that's not always true so we created a new method to
specifically clear the results.